refactor(tui): extract types, viewport, and history from tui_app.rs (#209)
PR 1/3 of the tui_app.rs god-function refactor.

Pure mechanical extraction — no behavioral changes:
- tui_types.rs: TuiState, MenuContent, PromptMode, ProviderWizard,
  type aliases (Term, *Dropdown), constants (MIN/MAX_VIEWPORT_HEIGHT)
- tui_viewport.rs: draw_viewport(), render_menu(), terminal lifecycle
  (init/restore/reinit/maybe_resize), emit_above()
- tui_history.rs: load_history(), save_history(), history_file_path()

tui_app.rs: 1919 → 1509 lines (~410 lines extracted)
tui_commands.rs: updated to use shared Term type alias

All 9 koda-cli tests pass. Zero behavioral change.
